Extract option screens from course choice flow into distinct controllers
The CandidateInterface::CourseChoicesController is becoming long,
complex, and difficult to understand. It requires refactoring.

As a first step, extract the action pairs related to choosing various
attributes of a course into their own controllers. Namespace these under
a CourseChoices module.
