+++ app/controllers/candidate_interface/course_choices/provider_selection_controller.rb
@@ -1,0 +1,22 @@
+module CandidateInterface
+  module CourseChoices
+    class ProviderSelectionController < BaseController
+      def options_for_provider
+        @pick_provider = PickProviderForm.new
+      end
+
+      def pick_provider
+        @pick_provider = PickProviderForm.new(
+          provider_id: params.dig(:candidate_interface_pick_provider_form, :provider_id),
+        )
+        render :options_for_provider and return unless @pick_provider.valid?
+
+        if @pick_provider.courses_available?
+          redirect_to candidate_interface_course_choices_course_path(@pick_provider.provider_id)
+        else
+          redirect_to candidate_interface_course_choices_ucas_no_courses_path(@pick_provider.provider_id)
+        end
+      end
+    end
+  end
+end
