+++ app/views/candidate_interface/course_choices/course_selection/options_for_course.html.erb
