+++ app/views/candidate_interface/course_choices/site_selection/options_for_site.html.erb
