+++ app/views/candidate_interface/course_choices/study_mode_selection/options_for_study_mode.html.erb
